LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/align-content-wrap-003-expected.txt


PASS .flexbox 1 
PASS .flexbox 2 
PASS .flexbox 3 
PASS .flexbox 4 
PASS .flexbox 5 
PASS .flexbox 6 
PASS .flexbox 7 
PASS .flexbox 8 
PASS .flexbox 9 
PASS .flexbox 10 
PASS .flexbox 11 
PASS .flexbox 12 
PASS .flexbox 13 
PASS .flexbox 14 
PASS .flexbox 15 
PASS .flexbox 16 
PASS .flexbox 17 
PASS .flexbox 18 
FAIL .flexbox 19 assert_equals: 
<div data-expected-height="30" class="flexbox horizontal" style="align-content: space-between; height: 30px">
  <div data-offset-x="0" data-offset-y="0" data-expected-height="10"></div>
</div>
height expected 10 but got 30
FAIL .flexbox 20 assert_equals: 
<div data-expected-height="30" class="flexbox horizontal" style="align-content: space-around; height: 30px">
  <div data-offset-x="0" data-offset-y="10" data-expected-height="10"></div>
</div>
height expected 10 but got 30
FAIL .flexbox 21 assert_equals: 
<div data-expected-height="30" class="flexbox horizontal" style="align-content: space-evenly; height: 30px">
  <div data-offset-x="0" data-offset-y="10" data-expected-height="10"></div>
</div>
height expected 10 but got 30
PASS .flexbox 22 
PASS .flexbox 23 
PASS .flexbox 24 
PASS .flexbox 25 
PASS .flexbox 26 
PASS .flexbox 27 
PASS .flexbox 28 
PASS .flexbox 29 
PASS .flexbox 30 
PASS .flexbox 31 
PASS .flexbox 32 
PASS .flexbox 33 
PASS .flexbox 34 
PASS .flexbox 35 
PASS .flexbox 36 
PASS .flexbox 37 
PASS .flexbox 38 
PASS .flexbox 39 
PASS .flexbox 40 
FAIL .flexbox 41 assert_equals: 
<div data-expected-width="30" class="flexbox vertical-rl" style="align-content: space-between; width: 30px">
  <div data-offset-x="20" data-offset-y="0" data-expected-width="10"></div>
</div>
width expected 10 but got 30
FAIL .flexbox 42 assert_equals: 
<div data-expected-width="30" class="flexbox vertical-rl" style="align-content: space-around; width: 30px">
  <div data-offset-x="10" data-offset-y="0" data-expected-width="10"></div>
</div>
width expected 10 but got 30
FAIL .flexbox 43 assert_equals: 
<div data-expected-width="30" class="flexbox vertical-rl" style="align-content: space-evenly; width: 30px">
  <div data-offset-x="10" data-offset-y="0" data-expected-width="10"></div>
</div>
width expected 10 but got 30
PASS .flexbox 44 
Test to make sure that align-content works properly.
